opt: move TypedExpr conversion code
Move the `scalarToTypedExpr` code to a separate file.

Release note: None
