+++ pkg/sql/opt/build.go
@@ -499,150 +499,6 @@ func buildScalar(pexpr tree.TypedExpr, evalCtx *tree.EvalContext) (_ *Expr, err 
 	return buildCtx.buildScalar(pexpr), nil
 }
 
-var typedExprConvMap [numOperators]func(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr
-
-func init() {
-	// This code is not inline to avoid an initialization loop error (some of the
-	// functions depend on scalarToTypedExpr which depends on typedExprConvMap).
-	typedExprConvMap = [numOperators]func(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr{
-		constOp:    constOpToTypedExpr,
-		variableOp: variableOpToTypedExpr,
-
-		andOp: boolOpToTypedExpr,
-		orOp:  boolOpToTypedExpr,
-		notOp: boolOpToTypedExpr,
-
-		unaryPlusOp:       unaryOpToTypedExpr,
-		unaryMinusOp:      unaryOpToTypedExpr,
-		unaryComplementOp: unaryOpToTypedExpr,
-
-		eqOp:           comparisonOpToTypedExpr,
-		ltOp:           comparisonOpToTypedExpr,
-		gtOp:           comparisonOpToTypedExpr,
-		leOp:           comparisonOpToTypedExpr,
-		geOp:           comparisonOpToTypedExpr,
-		neOp:           comparisonOpToTypedExpr,
-		inOp:           comparisonOpToTypedExpr,
-		notInOp:        comparisonOpToTypedExpr,
-		likeOp:         comparisonOpToTypedExpr,
-		notLikeOp:      comparisonOpToTypedExpr,
-		iLikeOp:        comparisonOpToTypedExpr,
-		notILikeOp:     comparisonOpToTypedExpr,
-		similarToOp:    comparisonOpToTypedExpr,
-		notSimilarToOp: comparisonOpToTypedExpr,
-		regMatchOp:     comparisonOpToTypedExpr,
-		notRegMatchOp:  comparisonOpToTypedExpr,
-		regIMatchOp:    comparisonOpToTypedExpr,
-		notRegIMatchOp: comparisonOpToTypedExpr,
-		isOp:           comparisonOpToTypedExpr,
-		isNotOp:        comparisonOpToTypedExpr,
-		containsOp:     comparisonOpToTypedExpr,
-		containedByOp:  comparisonOpToTypedExpr,
-		anyOp:          comparisonOpToTypedExpr,
-		someOp:         comparisonOpToTypedExpr,
-		allOp:          comparisonOpToTypedExpr,
-
-		bitandOp:   binaryOpToTypedExpr,
-		bitorOp:    binaryOpToTypedExpr,
-		bitxorOp:   binaryOpToTypedExpr,
-		plusOp:     binaryOpToTypedExpr,
-		minusOp:    binaryOpToTypedExpr,
-		multOp:     binaryOpToTypedExpr,
-		divOp:      binaryOpToTypedExpr,
-		floorDivOp: binaryOpToTypedExpr,
-		modOp:      binaryOpToTypedExpr,
-		powOp:      binaryOpToTypedExpr,
-		concatOp:   binaryOpToTypedExpr,
-		lShiftOp:   binaryOpToTypedExpr,
-		rShiftOp:   binaryOpToTypedExpr,
-
-		tupleOp: tupleOpToTypedExpr,
-
-		unsupportedScalarOp: unsupportedScalarOpToTypedExpr,
-	}
-}
-
-func constOpToTypedExpr(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr {
-	return e.private.(tree.Datum)
-}
-
-func variableOpToTypedExpr(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr {
-	return ivh.IndexedVar(e.private.(*columnProps).index)
-}
-
-func boolOpToTypedExpr(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr {
-	switch e.op {
-	case andOp, orOp:
-		n := scalarToTypedExpr(e.children[0], ivh)
-		for _, child := range e.children[1:] {
-			m := scalarToTypedExpr(child, ivh)
-			if e.op == andOp {
-				n = tree.NewTypedAndExpr(n, m)
-			} else {
-				n = tree.NewTypedOrExpr(n, m)
-			}
-		}
-		return n
-
-	case notOp:
-		return tree.NewTypedNotExpr(scalarToTypedExpr(e.children[0], ivh))
-	default:
-		panic(fmt.Sprintf("invalid op %s", e.op))
-	}
-}
-
-func tupleOpToTypedExpr(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr {
-	if isTupleOfConstants(e) {
-		datums := make(tree.Datums, len(e.children))
-		for i, child := range e.children {
-			datums[i] = constOpToTypedExpr(child, ivh).(tree.Datum)
-		}
-		return tree.NewDTuple(datums...)
-	}
-	children := make([]tree.TypedExpr, len(e.children))
-	for i, child := range e.children {
-		children[i] = scalarToTypedExpr(child, ivh)
-	}
-	return tree.NewTypedTuple(children)
-}
-
-func unaryOpToTypedExpr(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr {
-	return tree.NewTypedUnaryExpr(
-		unaryOpReverseMap[e.op],
-		scalarToTypedExpr(e.children[0], ivh),
-		e.scalarProps.typ,
-	)
-}
-
-func comparisonOpToTypedExpr(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr {
-	return tree.NewTypedComparisonExprWithSubOp(
-		comparisonOpReverseMap[e.op],
-		comparisonOpReverseMap[e.subOperator],
-		scalarToTypedExpr(e.children[0], ivh),
-		scalarToTypedExpr(e.children[1], ivh),
-	)
-}
-
-func binaryOpToTypedExpr(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr {
-	return tree.NewTypedBinaryExpr(
-		binaryOpReverseMap[e.op],
-		scalarToTypedExpr(e.children[0], ivh),
-		scalarToTypedExpr(e.children[1], ivh),
-		e.scalarProps.typ,
-	)
-}
-
-func unsupportedScalarOpToTypedExpr(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr {
-	return e.private.(tree.TypedExpr)
-}
-
-func scalarToTypedExpr(e *Expr, ivh *tree.IndexedVarHelper) tree.TypedExpr {
-	if fn := typedExprConvMap[e.op]; fn != nil {
-		return fn(e, ivh)
-	}
-	panic(fmt.Sprintf("unsupported op %s", e.op))
-}
-
 // BuildScalarExpr converts a TypedExpr to a *Expr tree and normalizes it.
 func BuildScalarExpr(typedExpr tree.TypedExpr, evalCtx *tree.EvalContext) (*Expr, error) {
 	if typedExpr == nil {
